requirements.txt
websockets==11.0.3
asyncio==3.4.3
pytest==7.4.3
pytest-asyncio==0.21.1
numpy==1.24.3
scipy==1.11.4
pyzmq==25.1.1 